Reverted code with user-friendly backup repo init process
Duplicacy by design does not save credentials to a keyring file passwords
provided via environmental variables. "The passwords will be stored when
the backup command (or any other command apart from init or add) is
run for the first time. This means you need to make sure that you do that
first run interactively, i.e. not via a script".

Because of that it has been needed to revert code that introduced
user-friendly backup repository init process.

References:
- https://github.com/gilbertchen/duplicacy/issues/241
- https://forum.duplicacy.com/t/passwords-credentials-and-environment-variables/1094